Only use 48 bits for encoding timestamps and 32 bits for encoding thread IDs in RawEvent in order to make it smaller.

+++ analyzeme/src/testing_common.rs
@@ -121,7 +121,7 @@ fn pseudo_invocation<S: SerializationSink>(
         return;
     }
 
-    let thread_id = (random % 3) as u64;
+    let thread_id = (random % 3) as u32;
 
     let (event_kind, event_id) = event_ids[random % event_ids.len()];
